# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZQAI API",
    "description": "REST API for the ZQAI portal (news, events, managers, departments, compliance officer, reception schedule, file uploads and user/auth management). Backend: Express 5 + PostgreSQL.\n\nNOTE: BearerAuth is documented as the intended security scheme for state-changing and administrative operations, but the current implementation does NOT enforce any authentication middleware. The live scan is expected to reveal this as a broken-authentication / broken-function-level-authorization finding.",
    "version": "0.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development server" }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Authenticate a user and issue a JWT",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Invalid credentials" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Create a new user (should be admin-only)",
        "tags": ["Auth"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserWrapper" } } }
          },
          "400": { "description": "username, password and role are required" },
          "401": { "description": "Unauthorized" },
          "409": { "description": "User already exists" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/auth/users": {
      "get": {
        "operationId": "getUsers",
        "summary": "List all users (should be admin-only)",
        "tags": ["Auth"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/auth/users/{id}/password": {
      "patch": {
        "operationId": "changeUserPassword",
        "summary": "Change a user's password (should be admin/self-only)",
        "tags": ["Auth"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordChangeRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Password updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "400": { "description": "Password too short" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/auth/users/{id}": {
      "delete": {
        "operationId": "deleteUser",
        "summary": "Delete a user (should be admin-only)",
        "tags": ["Auth"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    },
...
    "/api/departments": {
      "get": {
        "operationId": "getDepartments",
        "summary": "List departments with nested divisions and sectors",
        "tags": ["Departments"],
        "responses": {
          "200": {
            "description": "List of departments",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Department" } } } }
          },
          "500": { "description": "Failed to fetch departments" }
        }
      }
    },
    "/api/departments/{id}": {
      "get": {
        "operationId": "getDepartmentById",
        "summary": "Get a single department with nested divisions and sectors",
        "tags": ["Departments"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Department",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Department" } } }
          },
          "404": { "description": "Department not found" },
          "500": { "description": "Failed to fetch department" }
        }
      }
    },
    "/api/compliance-officer": {
      "get": {
        "operationId": "getComplianceOfficer",
        "summary": "Get compliance officer info",
        "tags": ["ComplianceOfficer"],
        "responses": {
          "200": {
            "description": "Compliance officer",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComplianceOfficer" } } }
          },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "operationId": "updateComplianceOfficer",
        "summary": "Create or update compliance officer info (should be authenticated)",
        "tags": ["ComplianceOfficer"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComplianceOfficerRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. NOTE: not currently enforced by any server middleware."
      }
    },
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 1, "maxLength": 255, "example": "admin" },
          "password": { "type": "string", "format": "password", "minLength": 1, "maxLength": 255, "example": "secret" }
        }
      },
      "UserSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "role": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserSummary" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "role"],
        "properties": {
          "username": { "type": "string", "minLength": 1, "maxLength": 255 },
          "password": { "type": "string", "format": "password", "minLength": 6, "maxLength": 255 },
          "role": { "type": "string", "enum": ["admin", "reception", "compliance_officer", "manager"] },
          "name": { "type": "string", "maxLength": 255 },
          "surname": { "type": "string", "maxLength": 255 }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "name": { "type": "string" },
          "surname": { "type": "string" },
          "role": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "UserWrapper": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserSummary" }
        }
      },
      "PasswordChangeRequest": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "password": { "type": "string", "format": "password", "minLength": 6, "maxLength": 255 }
        }
      },
...
      "Sector": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "division_id": { "type": "integer", "format": "int64" },
          "name_ru": { "type": "string" },
          "name_kk": { "type": "string" },
          "head_full_name": { "type": "string", "nullable": true },
          "head_title_ru": { "type": "string", "nullable": true },
          "head_title_kk": { "type": "string", "nullable": true }
        }
      },
      "Division": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "department_id": { "type": "integer", "format": "int64" },
          "name_ru": { "type": "string" },
          "name_kk": { "type": "string" },
          "head_full_name": { "type": "string", "nullable": true },
          "head_title_ru": { "type": "string", "nullable": true },
          "head_title_kk": { "type": "string", "nullable": true },
          "sectors": { "type": "array", "items": { "$ref": "#/components/schemas/Sector" } }
        }
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name_ru": { "type": "string" },
          "name_kk": { "type": "string" },
          "director_full_name": { "type": "string", "nullable": true },
          "director_title_ru": { "type": "string", "nullable": true },
          "director_title_kk": { "type": "string", "nullable": true },
          "divisions": { "type": "array", "items": { "$ref": "#/components/schemas/Division" } }
        }
      },
      "ComplianceOfficer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "full_name": { "type": "string" },
          "position": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email", "nullable": true },
          "reception_schedule": { "type": "string", "nullable": true },
          "photo_url": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ComplianceOfficerRequest": {
        "type": "object",
        "required": ["full_name", "position"],
        "properties": {
          "full_name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "position": { "type": "string", "maxLength": 255 },
          "phone": { "type": "string", "maxLength": 50, "nullable": true },
          "email": { "type": "string", "format": "email", "nullable": true },
          "reception_schedule": { "type": "string", "nullable": true },
          "photo_url": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true }
        }
      },
...
    }
  }
}